split deployment strategies into their own files
